Implemented integration over GP hyperparameters
GP prediction is done using a list of sklearn.gaussian_process.GaussianProcess
objects. This makes it difficult or impossible to incorporate signal variance
(i.e. sigma^2_f) into the model. The scaling of the "nugget" may be incorrect,
and the log probability functions may be off due to the normalization done by
scikit-learn.