Fix some scaling issues in source, add notebook to demo feature
As the TODO's in bayesian_optimization.py indicate, there may still be a few
adjustments needed to make this feature work exactly the way it should. This
mostly has to do with the fact that the target mean and signal variance (amp2)
are sampled and should be used in the GP fit and prediction processes (i.e. set
normalize to False and manually pre-process data using sampled hyper values).
Even without this aspect, comparison.ipynb demonstrates that integrating over
hyperparameters offers advantages.